refactor: add directory for cli
