+++ src/create_github_project/__main__.py
@@ -1,8 +1,18 @@
-from create_github_project import cli
+import click
+
+from create_github_project.commands import build
+
+
+@click.group(help='Command line tool to create templated Git project.')
+def cli() -> None:
+    """CLI のエントリーポイント。
+    """
+    pass
 
 
 def main():
-    cli.cmd()
+    build(cli)
+    cli()
 
 
 if __name__ == "__main__":
