+++ src/create_github_project/commands/accounts/_list.py
@@ -1,0 +1,10 @@
+import click
+
+from create_github_project.accounts import Accounts
+
+
+@click.command(name='list', help='List GitHub account under managements.')
+def _list() -> None:
+    """管理下にある GitHub アカウントの一覧を出力する。
+    """
+    Accounts().dump_list()
