+++ src/create_github_project/commands/versions/const.py
@@ -1,0 +1,1 @@
+REPOSITORY = 'nkomiya/create-github-project'
